[mojo-lang] Improve QoI for materialization errors around calls
This improves the "cannot materialize" error for a call to suggest
(with a fixit) wrapping the entire call with `comptime` when one
of the arguments cannot be materialized, e.g. when "lo" is
unmaterializable we now generate:

```
ex.mojo:7:13: error: cannot materialize comptime value of type 'Layout' to runtime because it is not 'ImplicitlyCopyable'
    var s = lo.size()
            ^
ex.mojo:7:20: note: use 'comptime' to evaluate the entire call at 'comptime' and materialize its result
    var s = lo.size()
                   ^
```

The rewrite converts the code to `var s = comptime(lo.size())`.

This also adds the proposal text to the repro for historical interest.

MODULAR_ORIG_COMMIT_REV_ID: 1e9ce01b8ea33d915f3c95706072f79a7abbfb22