[mojo-lang] Initial support for `__next__` raising `StopIteration` (#74426)
This is the first step towards supporting iterators that can
raise StopIteration instead of using the "has_next" protocol.
This is a step forward towards it, but there is more to do:
this is not ready for general tire kicking yet.

MODULAR_ORIG_COMMIT_REV_ID: a08cf86af9f556601adead2f6839543aea2d8328