[DeviceContext] Add non-owning constructor to DeviceContext
# Background

The lifetime of a DeviceContext object is maintained by a reference
count in C++. The Mojo DeviceContext is just a wrapper that holds a
pointer to the C++ object, forwards all calls through the FFI, and
maintains an associated refcount.

Within the graph, the lifetime of each DeviceContext is managed by the
graph compiler. Mojo Kernels that work with a DeviceContext receive a
DeviceContextPtr argument, which represents a pointer to an existing C++
DeviceContext object. Dereferencing DeviceContextPtr calls a private
constructor of DeviceContext that simply initializes the internal
pointer and bypasses the regular constructor call.

# Problem

The documentation for DeviceContextPtr states that we do not increment
the reference count when dereferencing to create a DeviceContext object:

Note: When initializing a `DeviceContext` from a pointer, the refcount
is not
    incremented. This is considered safe because `get_device_context()`
is only used within kernels and the `DeviceContext` lifetime is managed
    by the graph compiler.

This is not accurate, we do call retain and release, which adds a small
amount of overhead when working with these objects.

    @doc_private
    fn __init__(out self, ctx_ptr: _DeviceContextPtr):
"""Create a Mojo DeviceContext from a pointer to an existing C++ object.
        """
        self._handle = ctx_ptr
        self._retain()

But AsyncRT_DeviceContext_release also calls checkPendingWork(), which
can invoke event handlers and do other work with the context. As a
result, seemingly cheap calls like getting the device ID can force an
expensive context switch and do arbitrary work.

# Changes

Add an "owning" to DeviceContext boolean to mark whether an instance
owns a reference count to the C++ object. When we create a DeviceContext
via the private constructor, owning is set to false. When we move or
copy a DeviceContext, propagate the owning flag. Objects with
owning=False should not call release upon destruction.
MODULAR_ORIG_COMMIT_REV_ID: 475842823ba91b537ad9f67550a87ba72610d99a