[Driver] Small tweaks to host buffer cache tests to document need for explicit syncs
It appears that explicit stream syncs are necessary to ensure that
allocated memory is returned to the buffer cache.
These syncs are necessary for correctness.

This is because recording an event on a empty stream and polling it does
not guarantee that the event is ready.
It may take a few micros.
As such, `enqueueEventHandler()` immedietely followed by
`checkPendingWork()` is not enough to guarantee that the memory is
returned.

MODULAR_ORIG_COMMIT_REV_ID: fc85e3903136fc71cb4aac33f141b2fa198bea0c