[mojo-lang] Rewrite for-each/while loop IR representation (#74521)
Mojo previously used a design where `lit.loop` contained a "condition"
"body" and "else" region.  The condition region yielded a bool result
controlling whether the body was entered or control would go to the
else block.

This patch replaces this design with a new approach: just have a "body"
and "else" block, and let the body decide whether an how it wants to
exit. A normal "break" would exit the loop entirely, skipping the
"else",
so we introduce a new "lit.loop.break.else" terminator to get to the
"else.

The advantage of this design is that it is much easier to work with,
supports more flexible control flow, more naturally matches the
scoping rules used by for-each loops, and is easier to read in the
IR.  A win all around.  It also unblocks throwing-next.

MODULAR_ORIG_COMMIT_REV_ID: e58f1477a2130439f42474e6b1a67534172961fe