[mojo-lang] Fix an autoparam related crash. (#74525)
When inserting autoparams for _'s, insert them after any infer-only
parameters that already exist.  For example given
`S[a: Int, //, b: T[_]]`, insert the _ parameter after a instead of
before it.

This fixes a crash on the testcase, and also keeps autoparam order
list consistent with the order of the arguments they exist (rather
than backwards).  This is a bit of a hack though, because we still
crash on a dependent parameter (filed as MOCO-3007).  We need bigger
IR changes to fix this more holistically.

MODULAR_ORIG_COMMIT_REV_ID: 774a79698727ef2c0cab3d832281cb857a8a047c