[mojo-stdlib] Change `Origin` "mut" parameter to require keyword
This moves the "mut" parameter of Origin to be a kw-only parameter,
which aligns it with UnsafePointer and Span's infer-only mut
parameter.  Origin's 'mut' parameter will soon be infer-only as well,
and this stages in a bunch of noisy-but-boring changes.

MODULAR_ORIG_COMMIT_REV_ID: fad8fc25270852f263179d65a821331d50eaa8f7