[mojo-lang] Reject use of Self parameters when redef in an init. (#74584)
This teaches mojo to disable parameters on Self when they're being
redefined in the result type of an initializer, e.g. things like
this can never be parameter inferred:

```mojo
struct BadResultParams[a: Int]:
  fn __init__(x: Int, out self: BadResultParams[Self.a + 1]):
     pass
```

because there is no "a" that can be both "a" and "a+1".  While there
are some situations that could be allowed, it is best to just narrow
the language.  This will make parameter inference's job incrementally
simpler since it knows that no uses of rebound Self parameters need
to be inferred early.

MODULAR_ORIG_COMMIT_REV_ID: 88773cc36aaf88af1ecea2218b580bdb6ef89037