[mojo-lang] Rework emission of '__comptime_assert' message expr (#74603)
The previous code was trying to force the string to a StaticString,
but this doesn't work for subexpressions like String - such a value
has a "ComptimeOrigin" origin (because the String lives at comptime)
not a Static constant origin.  This happens to get through parameter
inference incorrectly, but will be caught in the future.

MODULAR_ORIG_COMMIT_REV_ID: 97cedc36c6db252679a1bb7ddad18a7055b3a539