[mojo-stdlib] Move LegacyUnsafePointer.mut to be inferred
`Origin` will soon be carrying its `!lit.origin` as a parameter instead
of state.  This means it becomes an autoparam for containing types
like LegacyUnsafePointer.  For this to be possible, the mut parameter
that it also depends on must be inferred.  This stages this in.

MODULAR_ORIG_COMMIT_REV_ID: 4d51bf37abf2926421d3cd55c8cdb762f570a40c